% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{add_centrality_measures}
\alias{add_centrality_measures}
\title{Add centrality measures to graph as node attributes}
\usage{
add_centrality_measures(g)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}
}
\value{
An igraph graph object.
}
\description{
Adds degree, in-degree, out-degree, betweenness and closeness measures to graph as node attributes.
}
