% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{addAdditionalMeasures}
\alias{addAdditionalMeasures}
\title{Add degree, in-degree, out-degree, betweenness and closeness measures to graph as vertex attributes}
\usage{
addAdditionalMeasures(g)
}
\arguments{
\item{g}{as graphml object}
}
\value{
g as graphml object
}
\description{
Add degree, in-degree, out-degree, betweenness and closeness measures to graph as vertex attributes
}
\keyword{internal}
