% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addLine}
\alias{addLine}
\title{Simple text aggregation}
\usage{
addLine(lines, add_line, txt = FALSE)
}
\arguments{
\item{lines}{vector of text lines}

\item{add_line}{text line to add to lines}

\item{txt}{return lines as character string}
}
\value{
text lines as vector or character string
}
\description{
Simple text aggregation
}
\keyword{internal}
