% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{apply_feature_filter}
\alias{apply_feature_filter}
\title{Filter out graph nodes and edges from graph object that are isolates, multi edge or edge loops}
\usage{
apply_feature_filter(g, isolates = TRUE, multi_edge = TRUE, loops_edge = TRUE)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{isolates}{Logical. Include isolate nodes in graph. Default is \code{TRUE}.}

\item{multi_edge}{Logical. Include multiple edges between nodes in graph. Default is \code{TRUE}.}

\item{loops_edge}{Logical. Include node edge loops in graph. Default is \code{TRUE}.}
}
\value{
An igraph graph object.
}
\description{
This function removes isolate nodes, multiple edges between nodes and or node edge loops from a
graph.
}
\note{
Removing multiple edges or edge loops from a graph will simplify it and remove other edge attributes.
}
