% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{applyCategoricalFilters}
\alias{applyCategoricalFilters}
\title{Filter out graph vertices not in selected voson category attributes}
\usage{
applyCategoricalFilters(g, selected_category, selected_category_attr)
}
\arguments{
\item{g}{graph as graphml object}

\item{selected_category}{vertex category as character string}

\item{selected_category_attr}{list of vertex category attributs to filter out as character string}
}
\value{
g as graphml object
}
\description{
Filter out graph vertices not in selected voson category attributes
}
\keyword{internal}
