% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{applyComponentFilter}
\alias{applyComponentFilter}
\title{Filter out graph vertices not in component size range}
\usage{
applyComponentFilter(g, component_type, component_range)
}
\arguments{
\item{g}{as graphml object}

\item{component_type}{as character string}

\item{component_range}{min and max values of range as vector}
}
\value{
g as graphml object
}
\description{
Filter out graph vertices not in component size range
}
\keyword{internal}
