% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{applyGraphFilters}
\alias{applyGraphFilters}
\title{Filter out graph vertices and edges from graph object that are isolates, multi edge or edge loops}
\usage{
applyGraphFilters(g, isolates, multi_edge, loops_edge)
}
\arguments{
\item{g}{as graphml object}

\item{isolates}{include isolate vertices in graph as logical}

\item{multi_edge}{include multiple edges between vertices in graph as logical}

\item{loops_edge}{include vertex edge loops in graph as logical}
}
\value{
g as graphml object
}
\description{
Filter out graph vertices and edges from graph object that are isolates, multi edge or edge loops
}
