% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{applyPruneFilter}
\alias{applyPruneFilter}
\title{Prune nodes from graph by node id}
\usage{
applyPruneFilter(g, selected_prune_nodes)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{selected_prune_nodes}{List. Selected node ids to remove.}
}
\value{
An igraph graph object.
}
\description{
This function removes a list of nodes from the graph object by node id value.
}
