% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{applyPruneFilterX}
\alias{applyPruneFilterX}
\title{Filter out list of vertices from graphml object using vertex id value}
\usage{
applyPruneFilterX(g, selected_prune_verts)
}
\arguments{
\item{g}{graph as graphml object}

\item{selected_prune_verts}{selected vertex ids to filter out as list}
}
\value{
g as graphml object
}
\description{
Filter out list of vertices from graphml object using vertex id value
}
