% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_interface.R
\name{collectMastodonSearchData}
\alias{collectMastodonSearchData}
\title{Collect mastodon data}
\usage{
collectMastodonSearchData(hashtag, instance, local = FALSE, n = 100)
}
\arguments{
\item{hashtag}{Character string. Hashtag used to search for posts.}

\item{instance}{Character string. Instance to collect comments from.}

\item{local}{Logical. Collect posts from local or global timeline.}

\item{n}{Numeric. Number of posts to retrieve.}
}
\value{
A vosonSML mastodon dataframe.
}
\description{
This function is a wrapper for collecting mastodon search comments using \code{vosonSML::Collect}.
}
\keyword{internal}
