% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_interface.R
\name{collectMastodonThreadData}
\alias{collectMastodonThreadData}
\title{Collect mastodon data}
\usage{
collectMastodonThreadData(mastodon_url_list)
}
\arguments{
\item{mastodon_url_list}{Character vector. Thread urls to collect comments from.}
}
\value{
A vosonSML mastodon dataframe.
}
\description{
This function is a wrapper for collecting mastodon thread comments using \code{vosonSML::Collect}.
}
\keyword{internal}
