% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{collectRedditData}
\alias{collectRedditData}
\title{Wrapper for collecting reddit thread comments using the vosonSML package}
\usage{
collectRedditData(reddit_url_list)
}
\arguments{
\item{reddit_url_list}{vector of thread urls to collect comments from}
}
\value{
data as vosonSML reddit collection dataframe
}
\description{
Wrapper for collecting reddit thread comments using the vosonSML package
}
\keyword{internal}
