% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{collectTwitterData}
\alias{collectTwitterData}
\title{Wrapper for collecting tweets using the vosonSML package}
\usage{
collectTwitterData(token, search_term, search_type, tweet_count,
  include_retweets, retry_on_rate_limit, language, date_until, since_id,
  max_id)
}
\arguments{
\item{token}{twitter auth token}

\item{search_term}{twitter search term as character string}

\item{search_type}{search type as character string - mixed, recent, popular}

\item{tweet_count}{number of tweets to collect}

\item{include_retweets}{logical whether to include retweets in the results}

\item{retry_on_rate_limit}{logical whether to wait and retry when the twitter api rate limit is hit}

\item{language}{language code of tweets to collect as two character ISO 639-1 code}

\item{date_until}{date to collect tweets to as character string YYYY-MM-DD}

\item{since_id}{numeric id get results with an id more recent than this id}

\item{max_id}{numeric id get results with an id older than this id}
}
\value{
tweets as vosonSML Collect dataframe
}
\description{
Wrapper for collecting tweets using the vosonSML package
}
\keyword{internal}
