% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{collectYoutubeData}
\alias{collectYoutubeData}
\title{Wrapper for collecting youtube video comments using the vosonSML package}
\usage{
collectYoutubeData(youtube_api_key, youtube_video_id_list,
  youtube_max_comments)
}
\arguments{
\item{youtube_api_key}{youtube api key as character string}

\item{youtube_video_id_list}{vector of youtube video ids to collect comments from}

\item{youtube_max_comments}{maximum number of comments to collect}
}
\value{
data as vosonSML youtube collection dataframe
}
\description{
Wrapper for collecting youtube video comments using the vosonSML package
}
\keyword{internal}
