% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dtm_filename}
\alias{create_dtm_filename}
\title{Create a file name with system datetime prefix}
\usage{
create_dtm_filename(suffix, ext, clean = FALSE)
}
\arguments{
\item{suffix}{Character string. Name part of file name to append to datetime part.}

\item{ext}{Character string. File extension without the period. For example, \code{"txt"}.}

\item{clean}{Logical. Remove problematic file system characters from file name part. Default is \code{FALSE}.}
}
\value{
A unique datetime filename as character string.
}
\description{
This function uses the system date and time to create a unique file name.
}
\keyword{internal}
