% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createRedditActorNetwork}
\alias{createRedditActorNetwork}
\title{Create reddit actor networks}
\usage{
createRedditActorNetwork(data)
}
\arguments{
\item{data}{\pkg{vosonSML} reddit dataframe.}
}
\value{
Reddit actor networks as named list.
}
\description{
This function is a wrapper for creating a reddit actor networks using \code{vosonSML::Create}.
}
\keyword{internal}
