% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createRedditRequestUrl}
\alias{createRedditRequestUrl}
\title{Return reddit url appropriate for reddit API comments request}
\usage{
createRedditRequestUrl(url)
}
\arguments{
\item{url}{reddit thread url}
}
\value{
url as character string
}
\description{
Return reddit url appropriate for reddit API comments request
}
\keyword{internal}
