% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTokenId}
\alias{createTokenId}
\title{Create a twitter auth token id}
\usage{
createTokenId(cred)
}
\arguments{
\item{cred}{\pkg{vosonSML} twitter credential object.}
}
\value{
A token id as character string.
}
\description{
This function uses properties of the twitter credential object to create a unique token id.
}
\keyword{internal}
