% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterActorNetwork}
\alias{createTwitterActorNetwork}
\title{Create twitter actor networks}
\usage{
createTwitterActorNetwork(data)
}
\arguments{
\item{data}{\pkg{vosonSML} twitter dataframe.}
}
\value{
Twitter actor networks as named list.
}
\description{
This function is a wrapper for creating a twitter actor networks using \code{vosonSML::Create}.
}
\keyword{internal}
