% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterDevToken}
\alias{createTwitterDevToken}
\title{Create a auth token with twitter app dev keys}
\usage{
createTwitterDevToken(app_name, keys)
}
\arguments{
\item{app_name}{twitter app name}

\item{keys}{named list of app API keys}
}
\value{
cred
}
\description{
Create a auth token with twitter app dev keys
}
\keyword{internal}
