% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterWebToken}
\alias{createTwitterWebToken}
\title{Create a auth token with twitter app consumer keys and interactive web authorization}
\usage{
createTwitterWebToken(app_name, keys)
}
\arguments{
\item{app_name}{twitter app name}

\item{keys}{named list of app API keys}
}
\value{
cred
}
\description{
Create a auth token with twitter app consumer keys and interactive web authorization
}
\keyword{internal}
