% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createYoutubeNetwork}
\alias{createYoutubeNetwork}
\title{Create youtube actor networks}
\usage{
createYoutubeNetwork(data)
}
\arguments{
\item{data}{\pkg{vosonSML} youtube dataframe.}
}
\value{
Youtube actor networks as named list.
}
\description{
This function is a wrapper for creating a youtube actor networks using \code{vosonSML::Create}.
}
\keyword{internal}
