% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{dash}
\alias{dash}
\title{Run the VOSON Dashboard Shiny Application}
\usage{
dash(startup_msgs = FALSE, local = NULL, ...)
}
\arguments{
\item{startup_msgs}{Logical. Display app package loading messages. Default is \code{FALSE}.}

\item{local}{Logical. Manually set app local or server mode flag.}

\item{...}{
  Arguments passed on to \code{\link[shiny:runApp]{shiny::runApp}}
  \describe{
    \item{\code{launch.browser}}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. The value of this parameter can also be a
function to call with the application's URL.}
    \item{\code{display.mode}}{The mode in which to display the application. If set to
the value \code{"showcase"}, shows application code and metadata from a
\code{DESCRIPTION} file in the application directory alongside the
application. If set to \code{"normal"}, displays the application normally.
Defaults to \code{"auto"}, which displays the application in the mode given
in its \code{DESCRIPTION} file, if any.}
  }}
}
\value{
None
}
\description{
This function launches the \pkg{VOSONDash} Shiny app in the default web browser.
}
