% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{dtbGraph}
\alias{dtbGraph}
\title{Load the package included "Divided They Blog" network graph}
\usage{
dtbGraph()
}
\value{
An igraph graph object.
}
\description{
This is a convenience function to load the "DividedTheyBlog_40Alist_release.graphml" graph.
}
\examples{
\dontrun{
# load the "Divided They Blog" network graph
g <- dtbGraph()
}

}
\keyword{internal}
