% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{eawGraph}
\alias{eawGraph}
\title{Load the package included "Enviro Activist Websites 2006" network graph}
\usage{
eawGraph()
}
\value{
An igraph graph object.
}
\description{
This is a convenience function to load the "enviroActivistWebsites_2006.graphml" graph.
}
\examples{
\dontrun{
# load the "Enviro Activist Websites 2006" network graph
g <- eawGraph()
}

}
\keyword{internal}
