% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{emptyPlotMessage}
\alias{emptyPlotMessage}
\title{Create an empty plot with text}
\usage{
emptyPlotMessage(message)
}
\arguments{
\item{message}{text to place on plot character string}
}
\value{
plot with text message
}
\description{
Create an empty plot with text
}
\keyword{internal}
