% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{filter_cats}
\alias{filter_cats}
\title{Filter out graph nodes not in selected category}
\usage{
filter_cats(g, selected_cat, selected_subcats, cat_prefix = "vosonCA_")
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{selected_cat}{Character string. Selected node category without prefix.}

\item{selected_subcats}{List. Selected sub-category values to include in graph.}

\item{cat_prefix}{Character string. Category attribute prefix format to match. Default is \code{"vosonCA_"}.}
}
\value{
An igraph graph object.
}
\description{
This function removes nodes that are not in the selected categories values list or sub-categories.
}
\examples{
\dontrun{
# return a graph containing only nodes that have the node category 
# attribute "vosonCA_Stance" value "liberal"
g <- get_pkg_data("DividedTheyBlog_40Alist_release.graphml")

g <- filter_cats(g, "Stance", c("liberal"))
}

}
