% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{filter_comps}
\alias{filter_comps}
\title{Filter out graph nodes not in component size range}
\usage{
filter_comps(g, mode = "strong", component_range)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{mode}{Character string. Use strongly or weakly connected components by specifying \code{"strong"} or
\code{"weak"}. Ignored for undirected graphs. Default is \code{"strong"}.}

\item{component_range}{Numeric vector. Min and max values or size range of component.}
}
\value{
An igraph graph object.
}
\description{
This function removes any graph nodes that are in components that fall outside of the specified
component size range.
}
