% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{filter_edges}
\alias{filter_edges}
\title{Filter out graph nodes and edges from graph object that are isolates, multi edge or edge loops}
\usage{
filter_edges(g, multi_edges = TRUE, loop_edges = TRUE)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{multi_edges}{Logical. Include multiple edges between nodes in graph. Default is \code{TRUE}.}

\item{loop_edges}{Logical. Include node edge loops in graph. Default is \code{TRUE}.}
}
\value{
An igraph graph object.
}
\description{
This function removes multiple edges between nodes and or node edge loops from a
graph.
}
\note{
Removing multiple edges or edge loops from a graph will simplify it and remove other edge attributes.
}
