% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{filter_nodes}
\alias{filter_nodes}
\title{Remove nodes from graph by node id}
\usage{
filter_nodes(g, ids = NULL, rm_iso = FALSE)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{ids}{List. Selected node ids to remove.}

\item{rm_iso}{Logical. Remove isolate nodes from graph (degree == 0).}
}
\value{
An igraph graph object.
}
\description{
This function removes a list of nodes from the graph object by node id value.
}
