% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{filter_nodes_iso}
\alias{filter_nodes_iso}
\title{Remove isolate nodes from graph}
\usage{
filter_nodes_iso(g, loops = FALSE)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{loops}{Logical. Whether to include loop edges in node degree calculation.}
}
\value{
An igraph graph object.
}
\description{
This function removes isolate nodes from the graph object.
}
