% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_attrs.R
\name{get_attr_properties}
\alias{get_attr_properties}
\title{Get properties of node and edge attributes}
\usage{
get_attr_properties(g, type = NULL, cat_max = NULL)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{type}{Character vector. Can contain "node" or "edge" as types. Default is \code{NULL} for both.}

\item{cat_max}{Numeric. Maximum number of unique values an attribute can have to be considered a category. By default
a category must have >= 1 and <= 25\% of its total values unique. Default is \code{NULL}.}
}
\value{
A tibble of attribute properties.
}
\description{
This function returns detected categorical, continuous, text and url type attributes.
}
