% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{get_comps_range}
\alias{get_comps_range}
\title{Get component count and min max size ranges for component modes}
\usage{
get_comps_range(g, mode = NULL)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{mode}{Character string. Use strongly or weakly connected components by specifying \code{"strong"} or
\code{"weak"}. Ignored for undirected graphs. Default is \code{"NULL"} for both modes.}
}
\value{
A named list of size and range values per component mode.
}
\description{
This function gets the number of components and the minimum size and maximum size range for all
components for a given mode.
}
