% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{get_dtb_graph}
\alias{get_dtb_graph}
\title{Load the package included "Divided They Blog" network graph}
\usage{
get_dtb_graph()
}
\value{
An igraph graph object.
}
\description{
This is a convenience function to load the "DividedTheyBlog_40Alist_release.graphml" graph.
}
\examples{
\dontrun{
# load the "Divided They Blog" network graph
g <- get_dtb_graph()
}

}
\keyword{internal}
