% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_eaw_graph}
\alias{get_eaw_graph}
\title{Load the package included "Enviro Activist Websites 2006" network graph}
\usage{
get_eaw_graph()
}
\value{
An igraph graph object.
}
\description{
This is a convenience function to load the "enviroActivistWebsites_2006.graphml" graph.
}
\examples{
\dontrun{
# load the "Enviro Activist Websites 2006" network graph
g <- get_eaw_graph()
}

}
\keyword{internal}
