% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{get_graph_centrlzn}
\alias{get_graph_centrlzn}
\title{Get graph centralization measures for an igraph object}
\usage{
get_graph_centrlzn(
  g,
  loops = TRUE,
  normalized = TRUE,
  directed = TRUE,
  warnings = FALSE
)
}
\arguments{
\item{g}{\pkg{igraph} graph object.}

\item{loops}{Logical; whether the loop edges are also counted.}

\item{normalized}{Logical scalar, whether to normalize the degree.  If
\code{TRUE} then the result is divided by \eqn{n-1}, where \eqn{n} is the
number of vertices in the graph.}

\item{directed}{logical scalar, whether to use directed shortest paths for
calculating betweenness.}

\item{warnings}{Logical. Suppress any warning messages. Default is \code{"TRUE"}.}
}
\value{
Graph centralization measures as named list
}
\description{
Function creates a list of centralization measures for a graph.
}
\details{
Default values for loops, normalized and directed are \code{"TRUE"}.
}
