% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{get_graph_metrics}
\alias{get_graph_metrics}
\title{Get network metrics from an igraph object}
\usage{
get_graph_metrics(g, mode = "strong", warnings = FALSE)
}
\arguments{
\item{g}{\pkg{igraph} graph object.}

\item{mode}{Character string, either \dQuote{weak} or \dQuote{strong}.  For
directed graphs \dQuote{weak} implies weakly, \dQuote{strong} strongly
connected components to search. It is ignored for undirected graphs.}

\item{warnings}{Logical. Suppress any warning messages. Default is \code{"TRUE"}.}
}
\value{
Network metrics as named list
}
\description{
Function creates a list of network metrics for a graph.
}
\details{
Default mode is \code{"strong"}.
}
