% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_mtdn.R
\name{get_mastodon_data}
\alias{get_mastodon_data}
\title{Collect mastodon data}
\usage{
get_mastodon_data(urls)
}
\arguments{
\item{urls}{Character vector. Thread urls to collect comments from.}
}
\value{
A vosonSML mastodon dataframe.
}
\description{
This function is a wrapper for collecting mastodon thread comments using \code{vosonSML::Collect}.
}
\keyword{internal}
