% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{get_node_cats}
\alias{get_node_cats}
\title{Get a list of node category attribute names and values}
\usage{
get_node_cats(g, cat_prefix = "vosonCA_", cat_colors = FALSE)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{cat_prefix}{Character string. Category attribute prefix format to match. Default is \code{"vosonCA_"}.}
}
\value{
A named list of node category attributes and values.
}
\description{
This function returns a list of graph node attribute names that match a category attribute prefix
format and their unique values.
}
\examples{
\dontrun{
# get a list of voson node categories and values
g <- get_pkg_data("DividedTheyBlog_40Alist_release.graphml")

cats <- get_node_cats(g)

# cats
# $Stance
# [1] "conservative" "liberal"  
}

}
