% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_cats.R
\name{get_node_properties}
\alias{get_node_properties}
\title{Get node attribute properties}
\usage{
get_node_properties(g)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}
}
\value{
A tibble of detected node attribute properties.
}
\description{
This function returns detected categorical, continuous, text and url type attributes.
}
