% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{get_pkg_data}
\alias{get_pkg_data}
\title{Load package included network graph}
\usage{
get_pkg_data(fname)
}
\arguments{
\item{fname}{Character string. Name of demonstration \code{graphml} file.}
}
\value{
An igraph graph object.
}
\description{
This function loads a network graph included in the \code{extdata} directory of the
\code{VOSONDash} package by file name.
}
\examples{
\dontrun{
# load the "Divided They Blog" package included network graph by file name
g <- get_pkg_data("DividedTheyBlog_40Alist_release.graphml")
}

}
