% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_pkg_data_list}
\alias{get_pkg_data_list}
\title{Return a list of graphml files found in package extdata}
\usage{
get_pkg_data_list()
}
\value{
A named list of file names.
}
\description{
This function returns a list of file names for graphs included in the \code{extdata} directory of the
\code{VOSONDash} package.
}
