% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_rddt.R
\name{get_rddt_data}
\alias{get_rddt_data}
\title{Get reddit thread data}
\usage{
get_rddt_data(urls = NULL, sort = NA, ...)
}
\arguments{
\item{urls}{Character vector. Thread urls to collect comments from..}

\item{...}{
  Arguments passed on to \code{\link[vosonSML:Collect.thread.reddit]{vosonSML::collect_reddit_threads}}
  \describe{
    \item{\code{sort}}{Character vector. Reddit comment sort order. Options are \code{"best"}, \code{"top"}, \code{"new"},
\code{"controversial"}, \code{"old"}, and \code{"qa"}. Default is \code{NA}.}
    \item{\code{waitTime}}{Numeric vector. Time range in seconds to select random wait from in-between url collection requests.
Minimum is 3 seconds. Default is \code{c(6, 8)} for a wait time chosen from between 6 and 8 seconds.}
    \item{\code{ua}}{Character string. Override User-Agent string to use in Reddit thread requests. Default is
\code{option("HTTPUserAgent")} value as set by vosonSML.}
  }}
}
\value{
A vosonSML reddit dataframe.
}
\description{
This function is a wrapper for collecting reddit thread comments using \code{vosonSML::Collect}.
}
