% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_rddt.R
\name{get_rddt_lst_data}
\alias{get_rddt_lst_data}
\title{Get reddit thread listings for one or more subreddits}
\usage{
get_rddt_lst_data(
  subreddits = NULL,
  sort = "new",
  period = "all",
  max = 25,
  ...
)
}
\arguments{
\item{subreddits}{Character vector. Subreddit urls to collect threads from.}

\item{...}{
  Arguments passed on to \code{\link[vosonSML:Collect.listing.reddit]{vosonSML::collect_reddit_listings}}
  \describe{
    \item{\code{sort}}{Character vector. Listing thread sort order. Options are \code{"hot"}, \code{"top"}, \code{"new"}, and
\code{"rising"}. Default is \code{"hot"}.}
    \item{\code{period}}{Character vector. Listing top threads by time period. Only applicable to sort order by \code{"top"}.
Options are \code{"hour"}, \code{"day"}, \code{"week"}, \code{"month"}, \code{"year"} and \code{"all"}. Default is
\code{"all"}.}
    \item{\code{max}}{Numeric vector. Maximum number of threads in listing to return. Default is \code{25}.}
    \item{\code{waitTime}}{Numeric vector. Time range in seconds to select random wait from in-between url collection requests.
Minimum is 3 seconds. Default is \code{c(6, 8)} for a wait time chosen from between 6 and 8 seconds.}
    \item{\code{ua}}{Character string. Override User-Agent string to use in Reddit thread requests. Default is
\code{option("HTTPUserAgent")} value as set by vosonSML.}
  }}
}
\value{
A vosonSML subreddit listing dataframe.
}
\description{
This function is a wrapper for collecting reddit subreddit listings using \code{vosonSML::Collect}.
}
