% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_rddt.R
\name{get_rddt_url_id}
\alias{get_rddt_url_id}
\title{Get a reddit thread id from url}
\usage{
get_rddt_url_id(url)
}
\arguments{
\item{url}{Character string. Reddit thread url.}
}
\value{
Reddit thread id as character string.
}
\description{
This function extracts the thread id from a reddit thread url.
}
