% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_rddt.R
\name{get_rddt_url_subrddt}
\alias{get_rddt_url_subrddt}
\title{Get subreddit name from url}
\usage{
get_rddt_url_subrddt(url)
}
\arguments{
\item{url}{Character string. Reddit thread url.}
}
\value{
Subreddit name as character string.
}
\description{
This function extracts the subreddit name from a reddit thread url.
}
