% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_interface.R
\name{get_reddit_data}
\alias{get_reddit_data}
\title{Collect reddit data}
\usage{
get_reddit_data(thread_urls = NULL, sort_type = NULL)
}
\arguments{
\item{thread_urls}{Character vector. Thread urls to collect comments from.}

\item{sort_type}{Character vector. Sort type for comments in each thread.}
}
\value{
A vosonSML reddit dataframe.
}
\description{
This function is a wrapper for collecting reddit thread comments using \code{vosonSML::Collect}.
}
\keyword{internal}
