% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sysfont_names}
\alias{get_sysfont_names}
\title{Check system fonts}
\usage{
get_sysfont_names()
}
\value{
Returns a character vector of installed system fonts.
}
\description{
Looks up installed system fonts.
}
\keyword{internal}
