% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_interface.R
\name{get_voson_ver}
\alias{get_voson_ver}
\title{Get the vosonSML package version}
\usage{
get_voson_ver()
}
\value{
Package version as character string.
}
\description{
This function returns the version of the loaded vosonSML package.
}
\keyword{internal}
