% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_voson_vers}
\alias{get_voson_vers}
\title{Get loaded voson package versions}
\usage{
get_voson_vers()
}
\value{
Package versions as named list.
}
\description{
This function returns the version of the vosonSML and VOSONDash packages.
}
\keyword{internal}
