% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_freqs.R
\name{get_word_freqs}
\alias{get_word_freqs}
\title{Get word frequencies}
\usage{
get_word_freqs(data, min = 1, top_n = NULL)
}
\arguments{
\item{data}{Dataframe. Word frequencies.}

\item{min}{Numeric. Minimum frequency for a word to be included in the results. Default is \code{1}.}

\item{top_n}{Numeric. Top number of words to return. Default is \code{20}.}
}
\value{
A dataframe of factored words.
}
\description{
Calculates word frequencies for a list or dataframe of words. Sets words as factors for top down highest
to lowest word frequency.
}
