% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_interface.R
\name{get_youtube_data}
\alias{get_youtube_data}
\title{Collect youtube data}
\usage{
get_youtube_data(api_key = NULL, video_ids = NULL, max_comments = NULL)
}
\arguments{
\item{api_key}{Character string. Youtube api key.}

\item{video_ids}{Character vector. Youtube video ids to collect comments from.}

\item{max_comments}{Numeric. Maximum number of comments to collect.}
}
\value{
A vosonSML youtube dataframe.
}
\description{
This function is a wrapper for collecting youtube video comments via \code{vosonSML::Collect}.
}
\keyword{internal}
