% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voson_ytbe.R
\name{get_ytbe_video_id}
\alias{get_ytbe_video_id}
\title{Get a youtube video id from url}
\usage{
get_ytbe_video_id(url)
}
\arguments{
\item{url}{Character string. Youtube video url.}
}
\value{
Video id as character string.
}
\description{
This function extracts the youtube video id from a youtube video url.
}
