% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{getColors}
\alias{getColors}
\title{Get the colors to use in a plot based on categorical attribute ordering applied to palette}
\usage{
getColors(categories, plot_category, plot_category_attrs, default_col,
  col_palette)
}
\arguments{
\item{categories}{list of all categorical attribures in the data set}

\item{plot_category}{selected plot category as character string}

\item{plot_category_attrs}{list of selected category attributes}

\item{default_col}{default color to use in plot}

\item{col_palette}{color palette to use in plot}
}
\value{
color values as character vector
}
\description{
Get the colors to use in a plot based on categorical attribute ordering applied to palette
}
\keyword{internal}
