% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkMetrics.R
\name{getNetworkMetrics}
\alias{getNetworkMetrics}
\title{Get graph network metrics}
\usage{
getNetworkMetrics(g, component_type = "strong")
}
\arguments{
\item{g}{\pkg{igraph} graph object.}

\item{component_type}{Character string. Use strongly or weakly connected components by specifying \code{"strong"} or 
\code{"weak"}. Ignored for undirected graphs. Default is \code{"strong"}.}
}
\value{
Network metrics as named vector.
}
\description{
Function creates a vector of calculated network metrics for a graph.
}
