% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{getNodeCategories}
\alias{getNodeCategories}
\title{Get a list of node category attribute names and values}
\usage{
getNodeCategories(g, cat_prefix = "vosonCA_")
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{cat_prefix}{Character string. Category attribute prefix format to match. Default is \code{"vosonCA_"}.}
}
\value{
A named list of node category attributes and values.
}
\description{
This function returns a list of graph node attribute names that match a category attribute prefix
format and their unique values.
}
\examples{
\dontrun{
# get a list of voson node categories and values
g <- loadPackageGraph("DividedTheyBlog_40Alist_release.graphml")

cats <- getNodeCategories(g)

# cats
# $Stance
# [1] "conservative" "liberal"  
}

}
