% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRedditUrlSubreddit}
\alias{getRedditUrlSubreddit}
\title{Get subreddit name from url}
\usage{
getRedditUrlSubreddit(url)
}
\arguments{
\item{url}{Character string. Reddit thread url.}
}
\value{
Subreddit name as character string.
}
\description{
This function extracts the subreddit name from a reddit thread url.
}
