% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRedditUrlThreadId}
\alias{getRedditUrlThreadId}
\title{Get a reddit thread id from url}
\usage{
getRedditUrlThreadId(url)
}
\arguments{
\item{url}{Character string. Reddit thread url.}
}
\value{
Reddit thread id as character string.
}
\description{
This function extracts the thread id from a reddit thread url.
}
