% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSystemFontFamilies}
\alias{getSystemFontFamilies}
\title{Check system fonts}
\usage{
getSystemFontFamilies()
}
\value{
Returns installed system font families.
}
\description{
Looks up installed system fonts.
}
\keyword{internal}
