% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{getVertexCategories}
\alias{getVertexCategories}
\title{Get a list of vertex attribute names that match a category attribute prefix format}
\usage{
getVertexCategories(g, cat_prefix = "vosonCA_")
}
\arguments{
\item{g}{as graphml object}

\item{cat_prefix}{character string of category attribute prefix format to match}
}
\value{
graph_cats
}
\description{
Get a list of vertex attribute names that match a category attribute prefix format
}
