% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getVOSONDashVer}
\alias{getVOSONDashVer}
\title{Get the VOSONDash package version}
\usage{
getVOSONDashVer()
}
\value{
Package version character string.
}
\description{
This function returns the version of the loaded VOSONDash package.
}
\keyword{internal}
