% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{getVosonSMLVersion}
\alias{getVosonSMLVersion}
\title{Get the vosonSML package version}
\usage{
getVosonSMLVersion(compare_ver)
}
\arguments{
\item{compare_ver}{Character string. Version string to compare.}
}
\value{
Package version as character string or if package later than compare version as logical.
}
\description{
This function returns the version of the loaded vosonSML package.
}
\keyword{internal}
