% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getYoutubeVideoId}
\alias{getYoutubeVideoId}
\title{Get a youtube video id from url}
\usage{
getYoutubeVideoId(url)
}
\arguments{
\item{url}{Character string. Youtube video url.}
}
\value{
Video id as character string.
}
\description{
This function extracts the youtube video id from a youtube video url.
}
