% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{hasVosonTextData}
\alias{hasVosonTextData}
\title{Check if graph object has vertex or edge voson text attributes}
\usage{
hasVosonTextData(g)
}
\arguments{
\item{g}{as graphml graph object}
}
\value{
has_text as boolean
}
\description{
Check if graph object has vertex or edge voson text attributes
}
\keyword{internal}
