% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{is_igraph}
\alias{is_igraph}
\title{Is an object an igraph object}
\usage{
is_igraph(g)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}
}
\value{
A logical value.
}
\description{
Test an object to see if it is an igraph object.
}
