% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNullOrEmpty}
\alias{isNullOrEmpty}
\title{Check a value for a range of empty conditions}
\usage{
isNullOrEmpty(x)
}
\arguments{
\item{x}{Input value.}
}
\value{
Result as logical.
}
\description{
This function checks a value is not NULL, NA or an empty string.
}
\keyword{internal}
