% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{isVosonSML0290}
\alias{isVosonSML0290}
\title{Return logical if vosonSML version later than 0.29}
\usage{
isVosonSML0290()
}
\value{
Logical.
}
\description{
This function returns if the installed version of vosonSML is later than v0.29
}
\keyword{internal}
