% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchVOSONDash.R
\name{launchVOSONDash}
\alias{launchVOSONDash}
\title{Launch the VOSON Dashboard Shiny Application}
\usage{
launchVOSONDash()
}
\value{
None
}
\description{
Launch the VOSON Dashboard Shiny Application
}
