% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{loadPackageGraph}
\alias{loadPackageGraph}
\title{Load package included network graph}
\usage{
loadPackageGraph(fname)
}
\arguments{
\item{fname}{Character string. Name of demonstration \code{graphml} file.}
}
\value{
An igraph graph object.
}
\description{
This function loads a network graph included in the \code{extdata} directory of the 
\code{VOSONDash} package by file name.
}
\examples{
\dontrun{
# load the "Divided They Blog" package included network graph by file name
g <- loadPackageGraph("DividedTheyBlog_40Alist_release.graphml")
}

}
