% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_queue}
\alias{log_queue}
\title{Add message to log queue}
\usage{
log_queue(queue, add_msg, txt = FALSE, max = 20)
}
\arguments{
\item{queue}{Character vector. Each item is a text log message.}

\item{add_msg}{Character string. Text log message to add to messages}

\item{txt}{Logical. Return messages as character string delimited with newline characters.}

\item{max}{Numeric. Return a message log queue of max messages, the rest are lost unless the queue was previously
queue was saved..}
}
\value{
Messages as vector or character string.
}
\description{
This function adds a text message to a queue or list with a count limiting how many messages are stored.
The queue stores max messages based on fifo.
}
\keyword{internal}
