% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logMessage}
\alias{logMessage}
\title{Simple logging queue with count FIFO}
\usage{
logMessage(messages, add_message, txt = FALSE, count = 20)
}
\arguments{
\item{messages}{vector of text messages}

\item{add_message}{text message to add to the messages queue}

\item{txt}{return lines as character string}

\item{count}{return queue of count messages and discard the rest}
}
\value{
messages as vector or character string
}
\description{
Simple logging queue with count FIFO
}
\keyword{internal}
