% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{mixmat}
\alias{mixmat}
\title{Create a mixing matrix}
\usage{
mixmat(g, attrib, use.density = TRUE)
}
\arguments{
\item{g}{graph as graphml object}

\item{attrib}{vertex attribute}

\item{use.density}{use edge density}
}
\value{
mixing matrix
}
\description{
mixing matrix original function written by Gary Weissman
see: https://gist.github.com/gweissman/2402741, http://www.babelgraph.org/wp/?p=351
}
\keyword{internal}
