% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVOSONDash.R
\name{runVOSONDash}
\alias{runVOSONDash}
\title{Run the VOSON Dashboard Shiny Application}
\usage{
runVOSONDash(pkgStartupMsgs = FALSE, isLocal = NULL)
}
\arguments{
\item{pkgStartupMsgs}{Logical. Display app package loading messages. Default is \code{FALSE}.}

\item{isLocal}{Logical. Manually set app local or server mode flag.}
}
\value{
None
}
\description{
This function launches the \pkg{VOSONDash} Shiny app in the default web browser.
}
