% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ids_and_labels}
\alias{set_ids_and_labels}
\title{Set graph ids and labels}
\usage{
set_ids_and_labels(g)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}
}
\value{
An igraph graph object.
}
\description{
Create simple node and edge ids and assign if missing. Rename existing id and label attributes as
imported. Will not change igraph symbolic names.
}
