% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{systemTimeFilename}
\alias{systemTimeFilename}
\title{Create a file-system friendly file name with system date time prefix}
\usage{
systemTimeFilename(name_suffix, name_ext, clean = FALSE)
}
\arguments{
\item{name_suffix}{name part of file name to append to date time as character string}

\item{name_ext}{file name extension as character string}

\item{clean}{remove problematic characters from file name as boolean}
}
\value{
file_name as character string
}
\description{
Create a file-system friendly file name with system date time prefix
}
\keyword{internal}
