% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VOSONDash-package.R
\docType{package}
\name{VOSONDash-package}
\alias{VOSONDash-package}
\alias{VOSONDash}
\title{Interface for collection and interactive analysis of social networks}
\description{
Description field.
}
\author{
Robert Ackland and Bryan Gertzel
}
