% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordCloudPlot}
\alias{wordCloudPlot}
\title{Creates a Word Cloud plot}
\usage{
wordCloudPlot(corp, seed, min_freq, max_words, pcolors = NULL)
}
\arguments{
\item{corp}{text corpus}

\item{seed}{value to seed rendering of word clouds}

\item{min_freq}{minimum word frequency for word clouds}

\item{max_words}{maximum number of words to render in word clouds}

\item{pcolors}{color list to use in the plot}
}
\value{
wordcloud plot
}
\description{
Creates a Word Cloud plot
}
\keyword{internal}
