% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordFreqChart}
\alias{wordFreqChart}
\title{Creates a Word Frequency chart}
\usage{
wordFreqChart(corp, min_freq, top_count, pcolors = NULL)
}
\arguments{
\item{corp}{text corpus}

\item{min_freq}{minimum word frequency for word frequency charts}

\item{top_count}{number of words to render in word frequency charts}

\item{pcolors}{color list to use in the plot}
}
\value{
barchart plot
}
\description{
Creates a Word Frequency chart
}
