% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordFreqFromCorpus}
\alias{wordFreqFromCorpus}
\title{Create a word frequency table from a text corpus}
\usage{
wordFreqFromCorpus(
  corp,
  rm_sparse = 0.99,
  word_len = c(8, 32),
  word_freq = c(1, Inf)
)
}
\arguments{
\item{corp}{a \pkg{tm} corpus object.}

\item{rm_sparse}{remove proportion sparse terms.}

\item{word_len}{word length filter.}

\item{word_freq}{min max word frequencies.}
}
\value{
A data.table of word frequencies.
}
\description{
Create a word frequency table.
}
