% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentChartSummary}
\alias{wordSentChartSummary}
\title{Creates a Sentiment Analysis chart}
\usage{
wordSentChartSummary(corp)
}
\arguments{
\item{corp}{text corpus}
}
\value{
barchart plot
}
\description{
Creates a Sentiment Analysis chart summary
}
