% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentData}
\alias{wordSentData}
\title{Create NRC Emotion data from words}
\usage{
wordSentData(corp)
}
\arguments{
\item{corp}{\pkg{tm} package document \code{\link[tm]{Corpus}} object.}
}
\value{
A nrc sentiment dataframe.
}
\description{
This function creates NRC data.
}
