% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{addAdditionalMeasures}
\alias{addAdditionalMeasures}
\title{Add additional measures to graph as vertex attributes}
\usage{
addAdditionalMeasures(g)
}
\arguments{
\item{g}{\pkg{igraph} \code{\link[igraph]{graph}} object.}
}
\value{
An igraph graph object.
}
\description{
Adds degree, in-degree, out-degree, betweenness and closeness measures to graph as vertex attributes.
}
