% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{applyPruneFilter}
\alias{applyPruneFilter}
\title{Prune vertices from graph by vertex id}
\usage{
applyPruneFilter(g, selected_prune_verts)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{selected_prune_verts}{List. Selected vertex ids to remove.}
}
\value{
An igraph graph object.
}
\description{
This function removes a list of vertices from the graph object by vertex id value.
}
