% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{collectRedditData}
\alias{collectRedditData}
\title{Collect reddit data}
\usage{
collectRedditData(reddit_url_list)
}
\arguments{
\item{reddit_url_list}{Character vector. Thread urls to collect comments from.}
}
\value{
A vosonSML reddit dataframe.
}
\description{
This function is a wrapper for collecting reddit thread comments using \code{vosonSML::Collect}.
}
\keyword{internal}
