% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{collectYoutubeData}
\alias{collectYoutubeData}
\title{Collect youtube data}
\usage{
collectYoutubeData(
  youtube_api_key,
  youtube_video_id_list,
  youtube_max_comments
)
}
\arguments{
\item{youtube_api_key}{Character string. Youtube api key.}

\item{youtube_video_id_list}{Character vector. Youtube video ids to collect comments from.}

\item{youtube_max_comments}{Numeric. Maximum number of comments to collect.}
}
\value{
A vosonSML youtube dataframe.
}
\description{
This function is a wrapper for collecting youtube video comments using \code{vosonSML::Collect}.
}
\keyword{internal}
