% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createRedditActorNetwork}
\alias{createRedditActorNetwork}
\title{Create reddit actor networks}
\usage{
createRedditActorNetwork(data)
}
\arguments{
\item{data}{as vosonSML reddit collection dataframe}
}
\value{
list of reddit actor networks as graphml objects
}
\description{
Create reddit actor networks
}
\keyword{internal}
