% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createRedditRequestUrl}
\alias{createRedditRequestUrl}
\title{Create a reddit url to request thread comments}
\usage{
createRedditRequestUrl(url)
}
\arguments{
\item{url}{Character string. Reddit thread url.}
}
\value{
Reddit API url as character string.
}
\description{
This function creates a url from specified thread url that can be used to request the thread comments.
}
\keyword{internal}
