% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTokenId}
\alias{createTokenId}
\title{Create voson dash twitter auth token id}
\usage{
createTokenId(token)
}
\arguments{
\item{token}{twitter auth token}
}
\value{
token_id as character string
}
\description{
Create voson dash twitter auth token id
}
\keyword{internal}
