% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterActorNetwork}
\alias{createTwitterActorNetwork}
\title{Create twitter actor networks}
\usage{
createTwitterActorNetwork(data)
}
\arguments{
\item{data}{vosonSML twitter Collect dataframe}
}
\value{
named list of twitter actor networks
}
\description{
Create twitter actor networks
}
\keyword{internal}
