% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterBearerToken}
\alias{createTwitterBearerToken}
\title{Create a auth token with twitter bearer token}
\usage{
createTwitterBearerToken(bearerToken)
}
\arguments{
\item{bearerToken}{Character string. Twitter app name.}
}
\value{
A \pkg{vosonSML} twitter credential object.
}
\description{
This function creates a \code{vosonSML::Authenticate} credential object with twitter bearer token.
}
\keyword{internal}
