% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterDevToken}
\alias{createTwitterDevToken}
\title{Create an auth token with twitter app dev keys}
\usage{
createTwitterDevToken(app_name, keys)
}
\arguments{
\item{app_name}{Character string. Twitter app name.}

\item{keys}{List. Named list of twitter app API keys.}
}
\value{
A vosonSML twitter credential object.
}
\description{
This function is a wrapper for \code{vosonSML::Authenticate} with twitter app developer keys. The 
properties \code{type} and \code{created} are added to the credential object to assist with \pkg{VOSONDash} token 
management.
}
\keyword{internal}
