% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createTwitterWebToken}
\alias{createTwitterWebToken}
\title{Create a auth token with twitter app consumer keys}
\usage{
createTwitterWebToken(app_name, keys)
}
\arguments{
\item{app_name}{Character string. Twitter app name.}

\item{keys}{List. Named list of twitter app API keys.}
}
\value{
A \pkg{vosonSML} twitter credential object.
}
\description{
This function creates a \code{vosonSML::Authenticate} credential object with twitter app consumer keys 
and interactive web authorization. \code{rtweet::create_token} is used to create the access token and the properties 
\code{type} and \code{created} are added to the credential object to assist with \pkg{VOSONDash} token management.
}
\keyword{internal}
