% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{createYoutubeNetwork}
\alias{createYoutubeNetwork}
\title{Create youtube actor networks}
\usage{
createYoutubeNetwork(data)
}
\arguments{
\item{data}{as vosonSML youtube collection dataframe}
}
\value{
named list of youtube actor networks
}
\description{
Create youtube actor networks
}
\keyword{internal}
