% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{emptyPlotMessage}
\alias{emptyPlotMessage}
\title{Create an empty plot with text}
\usage{
emptyPlotMessage(message = "No plot available.")
}
\arguments{
\item{message}{Character string. Text message to centre on empty plot. Default text is \code{"No plot available."}.}
}
\value{
An empty plot with text message.
}
\description{
This function creates an empty plot that can be used to display a text message. Intended to be used in 
a series of plots to indicate that an individual plot cannot be created for some reason and still maintain a plot 
aesthetic.
}
\keyword{internal}
