% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRedditUrlSubreddit}
\alias{getRedditUrlSubreddit}
\title{Extract subreddit name from a reddit thread url}
\usage{
getRedditUrlSubreddit(url)
}
\arguments{
\item{url}{reddit thread url}
}
\value{
subreddit as character string
}
\description{
Extract subreddit name from a reddit thread url
}
