% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRedditUrlThreadId}
\alias{getRedditUrlThreadId}
\title{Extract thread id from a reddit thread url}
\usage{
getRedditUrlThreadId(url)
}
\arguments{
\item{url}{reddit thread url}
}
\value{
thread_id as character string
}
\description{
Extract thread id from a reddit thread url
}
\keyword{internal}
