% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkGraphs.R
\name{getVertexCategories}
\alias{getVertexCategories}
\title{Get a list of vertex category attribute names}
\usage{
getVertexCategories(g, cat_prefix = "vosonCA_")
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}

\item{cat_prefix}{Character string. Category attribute prefix format to match. Default is \code{"vosonCA_"}.}
}
\value{
A list of vertex category attributes.
}
\description{
This function returns a list of graph vertex attribute names that match a category attribute prefix 
format.
}
