% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getVOSONDashVer}
\alias{getVOSONDashVer}
\title{Loaded VOSONDash package version}
\usage{
getVOSONDashVer()
}
\value{
package version as character string
}
\description{
Loaded VOSONDash package version
}
\keyword{internal}
