% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{getVosonSMLVersion}
\alias{getVosonSMLVersion}
\title{Get the vosonSML package version}
\usage{
getVosonSMLVersion()
}
\value{
Package version as character string.
}
\description{
This function returns the version of the loaded vosonSML package.
}
\keyword{internal}
