% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vosonSML.R
\name{getVosonSMLVersion}
\alias{getVosonSMLVersion}
\title{Loaded vosonSML package version}
\usage{
getVosonSMLVersion(compare_ver)
}
\value{
package version as character string
}
\description{
Loaded vosonSML package version
}
\keyword{internal}
