% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getYoutubeVideoId}
\alias{getYoutubeVideoId}
\title{Extract youtube video id from url}
\usage{
getYoutubeVideoId(url)
}
\arguments{
\item{url}{youtube video url}
}
\value{
video_id as character string
}
\description{
Extract youtube video id from url
}
