% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_filters.R
\name{hasVosonTextData}
\alias{hasVosonTextData}
\title{Check if graph object has text attributes}
\usage{
hasVosonTextData(g)
}
\arguments{
\item{g}{\pkg{igraph} \code{graph} object.}
}
\value{
Result as logical.
}
\description{
This function checks if a graph has either node or edge text attributes.
}
\note{
Uses the \code{VOSON} node and edge text attribute prefix \code{"vosonTxt_"} to determine if attributes are
text attributes.
}
\keyword{internal}
