% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNullOrEmpty}
\alias{isNullOrEmpty}
\title{Check a shiny app input value for a range of empty conditions}
\usage{
isNullOrEmpty(x)
}
\arguments{
\item{x}{shiny input value}
}
\value{
result as logical
}
\description{
Check a shiny app input value for a range of empty conditions
}
\keyword{internal}
