% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logMessage}
\alias{logMessage}
\title{Add message to log queue}
\usage{
logMessage(messages, add_message, txt = FALSE, count = 20)
}
\arguments{
\item{messages}{Character vector. Lines or text log messages.}

\item{add_message}{Character string. Text log message to add to messages}

\item{txt}{Logical. Return messages as single character string delimited by newline characters.}

\item{count}{Numeric. Return queue of count messages and discard the rest.}
}
\value{
Messages as vector or character string.
}
\description{
This function adds a text message to a queue or list with a count limiting how many messages are 
stored. The queue stores count messages based on first in first out.
}
\keyword{internal}
