% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{systemTimeFilename}
\alias{systemTimeFilename}
\title{Create a file name with system date time prefix}
\usage{
systemTimeFilename(name_suffix, name_ext, clean = FALSE)
}
\arguments{
\item{name_suffix}{Character string. Name part of file name to append to date time part.}

\item{name_ext}{Character string. File extension without the period. For example, \code{"txt"}.}

\item{clean}{Logical. Remove problematic file system characters from file name part. Default is \code{FALSE}.}
}
\value{
A unique date time file name as character string.
}
\description{
This function uses the system date and time to create a unique file name.
}
\keyword{internal}
