% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VOSONDash-package.R
\docType{package}
\name{VOSONDash-package}
\alias{VOSONDash-package}
\alias{VOSONDash}
\title{Interface for collection and interactive analysis of social networks}
\description{
VOSONDash provides functions and an interface in the form of an interactive R 
Shiny web application for the visualisation and analysis of network data. The app has 
sections for visualising and manipulating network graphs, performing text analysis, and 
displaying network metrics. It also has an interface for the collection of social network 
data using the vosonSML R package.
}
\author{
Bryan Gertzel and Robert Ackland.
}
