% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordCloudPlot}
\alias{wordCloudPlot}
\title{Creates a Word Cloud plot}
\usage{
wordCloudPlot(data, seed, categories, min_freq, max_words,
  col_palette = NULL)
}
\arguments{
\item{data}{voson dashboard text analysis plot list data structure
data structure is named list:
  list(graph_attr = list(attribute_name, attribute_value), VCorpus)}

\item{seed}{value to seed rendering of word clouds}

\item{categories}{list of all categorical attribures in the data set}

\item{min_freq}{minimum word frequency for word clouds}

\item{max_words}{maximum number of words to render in word clouds}

\item{col_palette}{color palette to use in the plot}
}
\value{
wordcloud
}
\description{
Creates a Word Cloud plot
}
