% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordCloudPlot}
\alias{wordCloudPlot}
\title{Create a Word Cloud plot}
\usage{
wordCloudPlot(corp, seed = NULL, min_freq = 1, max_words = 50, pcolors = NULL)
}
\arguments{
\item{corp}{\pkg{tm} package document \code{\link[tm]{Corpus}} object.}

\item{seed}{Numeric. Seed value can be supplied to reproduce a word cloud layout.}

\item{min_freq}{Numeric. Minimum word frequency to include a word in the word cloud. Default is \code{1}.}

\item{max_words}{Numeric. Maximum number of words to render in the word cloud. Default is \code{50}.}

\item{pcolors}{List. Colors to assign categorical variable in the plot. Default is \code{NULL}.}
}
\value{
A wordcloud plot.
}
\description{
This function creates a wordcloud plot of words in a text corpus.
}
