% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordFreqChart}
\alias{wordFreqChart}
\title{Creates a Word Frequency chart}
\usage{
wordFreqChart(data, categories, min_freq, top_count, col_palette = NULL)
}
\arguments{
\item{data}{voson dashboard text analysis plot list data structure
data structure is named list:
  list(graph_attr = list(attribute_name, attribute_value), VCorpus)}

\item{categories}{list of all categorical attribures in the data set}

\item{min_freq}{minimum word frequency for word frequency charts}

\item{top_count}{number of words to render in word frequency charts}

\item{col_palette}{color palette to use in the plot}
}
\value{
barchart plot
}
\description{
Creates a Word Frequency chart
}
