% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordFreqFromCorpus}
\alias{wordFreqFromCorpus}
\title{Create a word frequency dataframe}
\usage{
wordFreqFromCorpus(
  corp,
  rm_sparse = 0.99,
  word_len = c(3, 26),
  word_freq = c(1, Inf)
)
}
\arguments{
\item{corp}{a \pkg{tm} text corpus object.}

\item{rm_sparse}{Logical. Remove proportion of sparse terms. Default is \code{0.99}.}

\item{word_len}{Numeric vector. Min and max length of words to include. Default is \code{c(3, 26)}.}

\item{word_freq}{Numeric vector. Min and max frequency of words to include. Default is \code{c(1, Inf)}.}
}
\value{
A data.table of word frequencies.
}
\description{
Create a word frequency dataframe from a text corpus.
}
