% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentChart}
\alias{wordSentChart}
\title{Creates a Sentiment Analysis chart}
\usage{
wordSentChart(data, categories, col_palette = NULL)
}
\arguments{
\item{data}{voson dashboard text analysis plot list data structure
data structure is named list:
  list(graph_attr = list(attribute_name, attribute_value), VCorpus)}

\item{categories}{list of all categorical attribures in the data set}

\item{col_palette}{color palette to use in the plot}
}
\value{
barchart
}
\description{
Creates a Sentiment Analysis chart
}
