% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentChart}
\alias{wordSentChart}
\title{Creates a Sentiment Analysis chart}
\usage{
wordSentChart(corp, pcolors = NULL)
}
\arguments{
\item{corp}{text corpus}

\item{pcolors}{color list to use in the plot}
}
\value{
barchart plot
}
\description{
Creates a Sentiment Analysis chart
}
