% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentChart}
\alias{wordSentChart}
\title{Create an NRC emotion chart}
\usage{
wordSentChart(data, pcolors = NULL)
}
\arguments{
\item{data}{Dataframe. NRC emotions table.}

\item{pcolors}{List. Colors to assign categorical variable in the plot. Default is \code{NULL}.}
}
\value{
A barchart plot.
}
\description{
This function creates a horizontal barchart measuring and sorting the eight NRC lexicon emotions.
Emotions are measured as the proportion of the total value of the eight emotions in the text as a
percentage.
}
\note{
Uses the \pkg{syuzhet} package implementation of Saif Mohammad’s NRC Emotion lexicon.
}
