% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentData}
\alias{wordSentData}
\title{Create NRC emotion data}
\usage{
wordSentData(corp, word_len = c(3, 26))
}
\arguments{
\item{corp}{\pkg{tm} package document \code{\link[tm]{Corpus}} object.}

\item{word_len}{Numeric vector. Min and max length of words to include. Default is \code{c(3, 26)}.}
}
\value{
An NRC sentiment dataframe.
}
\description{
This function creates an NRC emotion dataframe from a text corpus.
}
\note{
Uses the \pkg{syuzhet} package implementation of Saif Mohammad’s NRC emotion lexicon.
}
