% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentValenceChart}
\alias{wordSentValenceChart}
\title{Creates a Sentiment Analysis chart}
\usage{
wordSentValenceChart(corp)
}
\arguments{
\item{corp}{text corpus}
}
\value{
barchart plot
}
\description{
Creates a Sentiment Analysis Valence chart
}
