% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentValenceChart}
\alias{wordSentValenceChart}
\title{Create an NRC sentiment valence chart}
\usage{
wordSentValenceChart(data)
}
\arguments{
\item{data}{Dataframe. NRC emotions table.}
}
\value{
A barchart plot.
}
\description{
This function creates a vertical barchart of the sum of negative and positive sentiments, and the
valence or net sentiment in a text corpus.
}
