% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{wordSentValenceChart}
\alias{wordSentValenceChart}
\title{Create an NRC Positive and Negative Sentiment valence chart}
\usage{
wordSentValenceChart(corp)
}
\arguments{
\item{corp}{\pkg{tm} package document \code{\link[tm]{Corpus}} object.}
}
\value{
A barchart plot.
}
\description{
This function creates a vertical barchart of negative, positive and the sum of the sentiment values or
valence in a text corpus.
}
\note{
Uses the \pkg{syuzhet} package implementation of Saif Mohammad’s NRC Emotion lexicon.
}
